Back to basics

+++ start-notebook.sh
@@ -14,8 +14,9 @@ HOME=$(eval echo "$HOME")
 JUPYTERHUB_USER=$REAL_JUPYTERHUB_USER # Swich back after expanding, as Jupyterhub breaks otherwise.
 
 mkdir -p "$HOME"
-#mv /home/notebook /home/oldnotebook
-ln -s "$HOME" /home/notebook
+if [ ! -L "/home/notebook" ]; then
+    ln -s "$HOME" /home/notebook
+fi
 
 # Exec the specified command or fall back on bash
 if [ $# -eq 0 ]; then
@@ -52,7 +53,6 @@ if [ -d "/mnt" ]; then
       fi
     done
 fi
-
 
 cd "$HOME"
 if [[ ! -z "${JUPYTER_ENABLE_LAB}" ]]; then
